requirements.txt
fastapi[standard]==0.135.2
google_api_python_client==2.193.0
pandas==3.0.2
protobuf==7.34.1
pydantic==2.12.5
python-dotenv==1.2.2
SQLAlchemy==2.0.30
uvicorn==0.42.0
google-genai==1.68.0
psycopg2-binary==2.9.11